Refactor tool tests into dedicated modules
