+++ tests/tools/address_tools/test_get_address_logs.py
