+++ tests/tools/transaction_tools/test_get_transactions_by_address.py
@@ -1,23 +1,11 @@
-# tests/tools/test_transaction_tools.py
 from unittest.mock import ANY, AsyncMock, MagicMock, patch
 
 import httpx
 import pytest
 
 from blockscout_mcp_server.config import config
-from blockscout_mcp_server.models import (
-    AdvancedFilterItem,
-    NextCallInfo,
-    PaginationInfo,
-    ToolResponse,
-    TransactionSummaryData,
-)
-from blockscout_mcp_server.tools.transaction_tools import (
-    get_token_transfers_by_address,
-    get_transaction_logs,
-    get_transactions_by_address,
-    transaction_summary,
-)
+from blockscout_mcp_server.models import AdvancedFilterItem, NextCallInfo, PaginationInfo, ToolResponse
+from blockscout_mcp_server.tools.transaction_tools import get_transactions_by_address
 
 
 @pytest.mark.asyncio
@@ -325,281 +313,6 @@ async def test_get_transactions_by_address_with_cursor_param(mock_ctx):
 
 
 @pytest.mark.asyncio
-async def test_get_token_transfers_by_address_calls_wrapper_correctly(mock_ctx):
-    """
-    Verify get_token_transfers_by_address calls the periodic progress wrapper with correct arguments.
-    """
-    # ARRANGE
-    chain_id = "1"
-    address = "0x123abc"
-    age_from = "2023-01-01T00:00:00.00Z"
-    age_to = "2023-01-02T00:00:00.00Z"
-    token = "0xA0b86a33E6441d95d7a9b6F25b0f2F5D6C16eD97"
-    mock_base_url = "https://eth.blockscout.com"
-    mock_api_response = {"items": [], "next_page_params": None}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
-        ) as mock_wrapper,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_wrapper.return_value = mock_api_response
-
-        # ACT
-        result = await get_token_transfers_by_address(
-            chain_id=chain_id, address=address, age_from=age_from, age_to=age_to, token=token, ctx=mock_ctx
-        )
-
-        # ASSERT
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, list)
-        assert result.data == []
-        mock_get_url.assert_called_once_with(chain_id)
-        mock_wrapper.assert_called_once()
-
-        # Check the wrapper call arguments
-        call_args, call_kwargs = mock_wrapper.call_args
-        assert call_kwargs["ctx"] == mock_ctx
-
-        from blockscout_mcp_server.tools.common import make_blockscout_request
-
-        assert call_kwargs["request_function"] == make_blockscout_request
-
-        # Check the request_args for token transfers
-        expected_request_args = {
-            "base_url": mock_base_url,
-            "api_path": "/api/v2/advanced-filters",
-            "params": {
-                "transaction_types": "ERC-20",
-                "to_address_hashes_to_include": address,
-                "from_address_hashes_to_include": address,
-                "age_from": age_from,
-                "age_to": age_to,
-                "token_contract_address_hashes_to_include": token,
-            },
-        }
-        assert call_kwargs["request_args"] == expected_request_args
-
-        # Verify other wrapper configuration
-        assert call_kwargs["tool_overall_total_steps"] == 2.0
-        assert call_kwargs["current_step_number"] == 2.0
-        assert call_kwargs["current_step_message_prefix"] == "Fetching token transfers"
-
-        # Verify progress was reported correctly before the wrapper call
-        assert mock_ctx.report_progress.call_count == 2
-
-
-@pytest.mark.asyncio
-async def test_get_token_transfers_by_address_chain_error(mock_ctx):
-    """
-    Verify that chain lookup errors are properly propagated without calling the wrapper.
-    """
-    # ARRANGE
-    chain_id = "999999"  # Invalid chain ID
-    address = "0x123abc"
-
-    from blockscout_mcp_server.tools.common import ChainNotFoundError
-
-    chain_error = ChainNotFoundError(f"Chain with ID '{chain_id}' not found on Chainscout.")
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
-        ) as mock_wrapper,
-    ):
-        mock_get_url.side_effect = chain_error
-
-        # ACT & ASSERT
-        with pytest.raises(ChainNotFoundError):
-            await get_token_transfers_by_address(chain_id=chain_id, address=address, ctx=mock_ctx)
-
-        # Verify the chain lookup was attempted
-        mock_get_url.assert_called_once_with(chain_id)
-
-        # Verify the wrapper was NOT called since chain lookup failed
-        mock_wrapper.assert_not_called()
-
-        # Progress should have been reported once (at start) before the error
-        assert mock_ctx.report_progress.call_count == 1
-        assert mock_ctx.info.call_count == 1
-
-
-@pytest.mark.asyncio
-async def test_get_token_transfers_by_address_transforms_response(mock_ctx):
-    """Verify that get_token_transfers_by_address correctly transforms its response."""
-    chain_id = "1"
-    address = "0x123"
-    mock_base_url = "https://eth.blockscout.com"
-
-    mock_api_response = {
-        "items": [
-            {
-                "from": {"hash": "0xfrom_hash"},
-                "to": {"hash": "0xto_hash"},
-                "token": "kept",
-                "total": "kept",
-                "value": "should be removed",
-                "internal_transaction_index": 1,
-                "created_contract": "should be removed",
-            }
-        ],
-        "next_page_params": None,
-    }
-
-    expected_items = [
-        {
-            "from": "0xfrom_hash",
-            "to": "0xto_hash",
-            "token": "kept",
-            "total": "kept",
-        }
-    ]
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
-        ) as mock_wrapper,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_wrapper.return_value = mock_api_response
-
-        result = await get_token_transfers_by_address(chain_id=chain_id, address=address, ctx=mock_ctx)
-
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, list)
-        assert len(result.data) == 1
-        item_model = result.data[0]
-        assert isinstance(item_model, AdvancedFilterItem)
-        assert item_model.from_address == expected_items[0]["from"]
-        assert item_model.to_address == expected_items[0]["to"]
-        item_dict = item_model.model_dump(by_alias=True)
-        assert item_dict["token"] == expected_items[0]["token"]
-        assert item_dict["total"] == expected_items[0]["total"]
-
-
-@pytest.mark.asyncio
-async def test_get_token_transfers_by_address_with_pagination(mock_ctx):
-    chain_id = "1"
-    address = "0x123abc"
-    mock_base_url = "https://eth.blockscout.com"
-
-    mock_api_response = {"items": []}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url",
-            new_callable=AsyncMock,
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress",
-            new_callable=AsyncMock,
-        ) as mock_wrapper,
-        patch("blockscout_mcp_server.tools.transaction_tools.create_items_pagination") as mock_create_pagination,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_wrapper.return_value = mock_api_response
-        mock_create_pagination.return_value = (
-            [],
-            PaginationInfo(
-                next_call=NextCallInfo(
-                    tool_name="get_token_transfers_by_address",
-                    params={"cursor": "CUR"},
-                )
-            ),
-        )
-
-        result = await get_token_transfers_by_address(chain_id=chain_id, address=address, ctx=mock_ctx)
-
-        mock_create_pagination.assert_called_once()
-        assert isinstance(result.pagination, PaginationInfo)
-
-
-@pytest.mark.asyncio
-async def test_get_token_transfers_by_address_custom_page_size(mock_ctx):
-    chain_id = "1"
-    address = "0x123"
-    mock_base_url = "https://eth.blockscout.com"
-
-    items = [{"block_number": i} for i in range(10)]
-    mock_api_response = {"items": items}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url",
-            new_callable=AsyncMock,
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress",
-            new_callable=AsyncMock,
-        ) as mock_wrapper,
-        patch("blockscout_mcp_server.tools.transaction_tools.create_items_pagination") as mock_create_pagination,
-        patch.object(config, "advanced_filters_page_size", 5),
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_wrapper.return_value = mock_api_response
-        mock_create_pagination.return_value = (items[:5], None)
-
-        await get_token_transfers_by_address(chain_id=chain_id, address=address, ctx=mock_ctx)
-
-        mock_create_pagination.assert_called_once()
-        assert mock_create_pagination.call_args.kwargs["page_size"] == 5
-
-
-@pytest.mark.asyncio
-async def test_get_token_transfers_by_address_with_cursor_param(mock_ctx):
-    chain_id = "1"
-    address = "0x123abc"
-    cursor = "CURSOR"
-    decoded = {"page": 2}
-    mock_base_url = "https://eth.blockscout.com"
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url",
-            new_callable=AsyncMock,
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress",
-            new_callable=AsyncMock,
-        ) as mock_wrapper,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.apply_cursor_to_params",
-        ) as mock_apply_cursor,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_wrapper.return_value = {"items": []}
-        mock_apply_cursor.side_effect = lambda cur, params: params.update(decoded)
-
-        await get_token_transfers_by_address(
-            chain_id=chain_id,
-            address=address,
-            cursor=cursor,
-            ctx=mock_ctx,
-        )
-
-        mock_apply_cursor.assert_called_once_with(cursor, ANY)
-        call_args, call_kwargs = mock_wrapper.call_args
-        params = call_kwargs["request_args"]["params"]
-        expected_params = {
-            "transaction_types": "ERC-20",
-            "to_address_hashes_to_include": address,
-            "from_address_hashes_to_include": address,
-            **decoded,
-        }
-        assert params == expected_params
-
-
-@pytest.mark.asyncio
 async def test_get_transactions_by_address_smart_pagination_error(mock_ctx):
     """
     Verify that errors from the smart pagination function are properly propagated.
@@ -841,159 +554,6 @@ async def test_get_transactions_by_address_multi_page_progress_reporting(mock_ct
         assert smart_pagination_call_args["total_steps"] == 12.0
         assert smart_pagination_call_args["ctx"] == mock_ctx
 
-        # Note: The smart pagination function internally handles progress reporting
-        # for steps 2-11 using make_request_with_periodic_progress for each page fetch
-
-
-@pytest.mark.asyncio
-async def test_transaction_summary_without_wrapper(mock_ctx):
-    """
-    Test a transaction tool that doesn't use the periodic progress wrapper for comparison.
-    This helps verify our testing approach for wrapper vs non-wrapper tools.
-    """
-    # ARRANGE
-    chain_id = "1"
-    tx_hash = "0x123abc"
-    mock_base_url = "https://eth.blockscout.com"
-
-    summary_obj = {"template": "This is a test transaction summary.", "vars": {}}
-    mock_api_response = {"data": {"summaries": [summary_obj]}}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_blockscout_request", new_callable=AsyncMock
-        ) as mock_request,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_request.return_value = mock_api_response
-
-        # ACT
-        result = await transaction_summary(chain_id=chain_id, transaction_hash=tx_hash, ctx=mock_ctx)
-
-        # ASSERT
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, TransactionSummaryData)
-        assert result.data.summary == [summary_obj]
-        mock_get_url.assert_called_once_with(chain_id)
-        mock_request.assert_called_once_with(base_url=mock_base_url, api_path=f"/api/v2/transactions/{tx_hash}/summary")
-
-        # This tool should have 3 progress reports (start, after URL, completion)
-        assert mock_ctx.report_progress.call_count == 3
-        assert mock_ctx.info.call_count == 3
-
-
-@pytest.mark.asyncio
-async def test_transaction_summary_no_summary_available(mock_ctx):
-    """
-    Test transaction_summary when no summary is available in the response.
-    """
-    # ARRANGE
-    chain_id = "1"
-    tx_hash = "0x123abc"
-    mock_base_url = "https://eth.blockscout.com"
-
-    # Response with no summary data
-    mock_api_response = {"data": {}}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_blockscout_request", new_callable=AsyncMock
-        ) as mock_request,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_request.return_value = mock_api_response
-
-        # ACT
-        result = await transaction_summary(chain_id=chain_id, transaction_hash=tx_hash, ctx=mock_ctx)
-
-        # ASSERT
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, TransactionSummaryData)
-        assert result.data.summary is None
-        mock_get_url.assert_called_once_with(chain_id)
-        mock_request.assert_called_once_with(base_url=mock_base_url, api_path=f"/api/v2/transactions/{tx_hash}/summary")
-        assert mock_ctx.report_progress.call_count == 3
-        assert mock_ctx.info.call_count == 3
-
-
-@pytest.mark.asyncio
-async def test_transaction_summary_handles_non_string_summary(mock_ctx):
-    """Verify transaction_summary correctly handles a non-string summary."""
-    # ARRANGE
-    chain_id = "1"
-    tx_hash = "0xcomplex"
-    mock_base_url = "https://eth.blockscout.com"
-
-    complex_summary = [
-        {"template": "Summary 1", "vars": {"a": 1}},
-        {"template": "Summary 2", "vars": {"b": 2}},
-    ]
-    mock_api_response = {"data": {"summaries": complex_summary}}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url",
-            new_callable=AsyncMock,
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_blockscout_request",
-            new_callable=AsyncMock,
-        ) as mock_request,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_request.return_value = mock_api_response
-
-        # ACT
-        result = await transaction_summary(chain_id=chain_id, transaction_hash=tx_hash, ctx=mock_ctx)
-
-        # ASSERT
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, TransactionSummaryData)
-        assert result.data.summary == complex_summary  # Assert it's the original list
-        mock_get_url.assert_called_once_with(chain_id)
-        mock_request.assert_called_once_with(base_url=mock_base_url, api_path=f"/api/v2/transactions/{tx_hash}/summary")
-        assert mock_ctx.report_progress.call_count == 3
-        assert mock_ctx.info.call_count == 3
-
-
-@pytest.mark.asyncio
-async def test_transaction_summary_handles_empty_list(mock_ctx):
-    """Return an empty list when Blockscout summarizes to nothing."""
-    chain_id = "1"
-    tx_hash = "0xempty"
-    mock_base_url = "https://eth.blockscout.com"
-
-    mock_api_response = {"data": {"summaries": []}}
-
-    with (
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url",
-            new_callable=AsyncMock,
-        ) as mock_get_url,
-        patch(
-            "blockscout_mcp_server.tools.transaction_tools.make_blockscout_request",
-            new_callable=AsyncMock,
-        ) as mock_request,
-    ):
-        mock_get_url.return_value = mock_base_url
-        mock_request.return_value = mock_api_response
-
-        result = await transaction_summary(chain_id=chain_id, transaction_hash=tx_hash, ctx=mock_ctx)
-
-        assert isinstance(result, ToolResponse)
-        assert isinstance(result.data, TransactionSummaryData)
-        assert result.data.summary == []
-        mock_get_url.assert_called_once_with(chain_id)
-        mock_request.assert_called_once_with(base_url=mock_base_url, api_path=f"/api/v2/transactions/{tx_hash}/summary")
-        assert mock_ctx.report_progress.call_count == 3
-        assert mock_ctx.info.call_count == 3
-
 
 @pytest.mark.asyncio
 async def test_get_transactions_by_address_invalid_cursor(mock_ctx):
@@ -1008,24 +568,280 @@ async def test_get_transactions_by_address_invalid_cursor(mock_ctx):
 
 
 @pytest.mark.asyncio
-async def test_get_token_transfers_by_address_invalid_cursor(mock_ctx):
-    """Verify ValueError is raised when the cursor is invalid."""
-    with patch(
-        "blockscout_mcp_server.tools.transaction_tools.apply_cursor_to_params",
-        side_effect=ValueError("invalid"),
-    ) as mock_apply:
-        with pytest.raises(ValueError, match="invalid"):
-            await get_token_transfers_by_address(chain_id="1", address="0xabc", cursor="bad", ctx=mock_ctx)
-    mock_apply.assert_called_once()
+async def test_get_transactions_by_address_multi_page_fetching(mock_ctx):
+    """
+    Test that get_transactions_by_address fetches multiple pages when initial results are sparse due to filtering.
+
+    This test simulates the scenario where each page contains mostly filtered-out transactions
+    (ERC-20, ERC-721, etc.) but some valid transactions, requiring multiple pages to accumulate
+    enough results for pagination decision.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    # Mock page 1: mostly filtered transactions with 2 valid ones
+    page1_items = [
+        {"type": "call", "hash": "0x1", "block_number": 100},
+        {"type": "ERC-20", "hash": "0x2", "block_number": 99},  # will be filtered out
+        {"type": "call", "hash": "0x3", "block_number": 98},
+        {"type": "ERC-721", "hash": "0x4", "block_number": 97},  # will be filtered out
+        {"type": "ERC-1155", "hash": "0x5", "block_number": 96},  # will be filtered out
+    ]
+
+    # Mock page 2: mostly filtered transactions with 2 valid ones
+    page2_items = [
+        {"type": "call", "hash": "0x6", "block_number": 95},
+        {"type": "ERC-20", "hash": "0x7", "block_number": 94},  # will be filtered out
+        {"type": "call", "hash": "0x8", "block_number": 93},
+        {"type": "ERC-404", "hash": "0x9", "block_number": 92},  # will be filtered out
+    ]
+
+    # Mock page 3: enough valid transactions to trigger pagination
+    page3_items = [
+        {"type": "call", "hash": "0x10", "block_number": 91},
+        {"type": "call", "hash": "0x11", "block_number": 90},
+        {"type": "call", "hash": "0x12", "block_number": 89},
+        {"type": "call", "hash": "0x13", "block_number": 88},
+        {"type": "call", "hash": "0x14", "block_number": 87},
+        {"type": "call", "hash": "0x15", "block_number": 86},
+        {"type": "call", "hash": "0x16", "block_number": 85},
+        {"type": "call", "hash": "0x17", "block_number": 84},
+        {"type": "call", "hash": "0x18", "block_number": 83},
+        {"type": "call", "hash": "0x19", "block_number": 82},
+    ]
+
+    # Expected API responses for each page
+    api_responses = [
+        {"items": page1_items, "next_page_params": {"page": 2}},
+        {"items": page2_items, "next_page_params": {"page": 3}},
+        {"items": page3_items, "next_page_params": {"page": 4}},
+    ]
+
+    with (
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
+        ) as mock_get_url,
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
+        ) as mock_request,
+        patch.object(config, "advanced_filters_page_size", 10),
+    ):
+        mock_get_url.return_value = mock_base_url
+        mock_request.side_effect = api_responses
+
+        # ACT
+        result = await get_transactions_by_address(
+            chain_id=chain_id,
+            address=address,
+            ctx=mock_ctx,
+        )
+
+        # ASSERT
+        # Should have called make_request_with_periodic_progress 3 times (3 pages)
+        assert mock_request.call_count == 3
+
+        # Should have accumulated filtered transactions from all 3 pages
+        # Page 1: 2 valid transactions (types "call")
+        # Page 2: 2 valid transactions (types "call")
+        # Page 3: 10 valid transactions (types "call")
+        # Total: 14 valid transactions
+        assert len(result.data) == 10  # Should be sliced to page_size
+
+        # Should have pagination since we have more than page_size valid transactions
+        assert result.pagination is not None
+
+        # Verify the transactions are properly transformed and ordered
+        assert all(item.type == "call" for item in result.data)
+        assert result.data[0].hash == "0x1"  # First transaction from page 1
+        assert result.data[1].hash == "0x3"  # Second valid transaction from page 1
 
 
 @pytest.mark.asyncio
-async def test_get_transaction_logs_invalid_cursor(mock_ctx):
-    """Verify ValueError is raised when the cursor is invalid."""
-    with patch(
-        "blockscout_mcp_server.tools.transaction_tools.apply_cursor_to_params",
-        side_effect=ValueError("bad"),
-    ) as mock_apply:
-        with pytest.raises(ValueError, match="bad"):
-            await get_transaction_logs(chain_id="1", transaction_hash="0xhash", cursor="bad", ctx=mock_ctx)
-    mock_apply.assert_called_once()
+async def test_get_transactions_by_address_stops_at_10_pages(mock_ctx):
+    """
+    Test that get_transactions_by_address stops fetching at 10 pages maximum.
+
+    This test ensures that even if there are more pages available, the function
+    will stop at 10 pages to prevent infinite loops or excessive API calls.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    # Create 10 pages of sparse results (1 valid transaction per page)
+    api_responses = []
+    for i in range(10):
+        page_items = [
+            {"type": "call", "hash": f"0x{i + 1}", "block_number": 100 - i},
+            {"type": "ERC-20", "hash": f"0x{i + 100}", "block_number": 99 - i},  # filtered out
+            {"type": "ERC-721", "hash": f"0x{i + 200}", "block_number": 98 - i},  # filtered out
+        ]
+        api_responses.append(
+            {
+                "items": page_items,
+                "next_page_params": {"page": i + 2},  # Always indicate more pages
+            }
+        )
+
+    with (
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
+        ) as mock_get_url,
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
+        ) as mock_request,
+        patch.object(
+            config, "advanced_filters_page_size", 20
+        ),  # Large page size to ensure we don't hit pagination limit
+    ):
+        mock_get_url.return_value = mock_base_url
+        mock_request.side_effect = api_responses
+
+        # ACT
+        result = await get_transactions_by_address(
+            chain_id=chain_id,
+            address=address,
+            ctx=mock_ctx,
+        )
+
+        # ASSERT
+        # Should have called make_request_with_periodic_progress exactly 10 times (max pages)
+        assert mock_request.call_count == 10
+
+        # Should have accumulated 10 valid transactions (1 per page)
+        assert len(result.data) == 10
+
+        # Should still have pagination since we stopped at max pages and had next_page_params
+        assert result.pagination is not None
+
+        # Verify all transactions are valid (not filtered out)
+        assert all(item.type == "call" for item in result.data)
+        assert result.data[0].hash == "0x1"
+        assert result.data[9].hash == "0x10"
+
+
+@pytest.mark.asyncio
+async def test_get_transactions_by_address_single_page_sufficient(mock_ctx):
+    """
+    Test that get_transactions_by_address works correctly when a single page has sufficient results.
+
+    This test ensures that when the first page contains enough valid transactions
+    after filtering, no additional pages are fetched.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    # Create a single page with enough valid transactions
+    page_items = [{"type": "call", "hash": f"0x{i + 1}", "block_number": 100 - i} for i in range(15)]
+
+    # Add some filtered transactions to ensure filtering works
+    page_items.extend(
+        [
+            {"type": "ERC-20", "hash": "0x100", "block_number": 85},
+            {"type": "ERC-721", "hash": "0x101", "block_number": 84},
+        ]
+    )
+
+    api_response = {
+        "items": page_items,
+        "next_page_params": {"page": 2},  # Indicate more pages available
+    }
+
+    with (
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
+        ) as mock_get_url,
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
+        ) as mock_request,
+        patch.object(config, "advanced_filters_page_size", 10),
+    ):
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = api_response
+
+        # ACT
+        result = await get_transactions_by_address(
+            chain_id=chain_id,
+            address=address,
+            ctx=mock_ctx,
+        )
+
+        # ASSERT
+        # Should have called make_request_with_periodic_progress only once
+        assert mock_request.call_count == 1
+
+        # Should have 10 transactions (page size limit)
+        assert len(result.data) == 10
+
+        # Should have pagination since we have more than page_size valid transactions
+        assert result.pagination is not None
+
+        # Verify all returned transactions are valid (not filtered out)
+        assert all(item.type == "call" for item in result.data)
+
+
+@pytest.mark.asyncio
+async def test_get_transactions_by_address_no_more_pages_available(mock_ctx):
+    """
+    Test that get_transactions_by_address correctly handles the case when no more pages are available.
+
+    This test ensures that when the API indicates no more pages (next_page_params is None),
+    the function returns all available results without pagination.
+    """
+    # ARRANGE
+    chain_id = "1"
+    address = "0x123abc"
+    mock_base_url = "https://eth.blockscout.com"
+
+    # Create a single page with few valid transactions and no next page
+    page_items = [
+        {"type": "call", "hash": "0x1", "block_number": 100},
+        {"type": "call", "hash": "0x2", "block_number": 99},
+        {"type": "ERC-20", "hash": "0x3", "block_number": 98},  # filtered out
+        {"type": "call", "hash": "0x4", "block_number": 97},
+    ]
+
+    api_response = {
+        "items": page_items,
+        "next_page_params": None,  # No more pages available
+    }
+
+    with (
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.get_blockscout_base_url", new_callable=AsyncMock
+        ) as mock_get_url,
+        patch(
+            "blockscout_mcp_server.tools.transaction_tools.make_request_with_periodic_progress", new_callable=AsyncMock
+        ) as mock_request,
+        patch.object(config, "advanced_filters_page_size", 10),
+    ):
+        mock_get_url.return_value = mock_base_url
+        mock_request.return_value = api_response
+
+        # ACT
+        result = await get_transactions_by_address(
+            chain_id=chain_id,
+            address=address,
+            ctx=mock_ctx,
+        )
+
+        # ASSERT
+        # Should have called make_request_with_periodic_progress only once
+        assert mock_request.call_count == 1
+
+        # Should have 3 valid transactions (filtered out 1 ERC-20)
+        assert len(result.data) == 3
+
+        # Should NOT have pagination since we have fewer than page_size and no more pages
+        assert result.pagination is None
+
+        # Verify all returned transactions are valid (not filtered out)
+        assert all(item.type == "call" for item in result.data)
+        assert result.data[0].hash == "0x1"
+        assert result.data[1].hash == "0x2"
+        assert result.data[2].hash == "0x4"
